+++ app/templates/_webpack-test.config.js
@@ -29,7 +29,7 @@ module.exports = {
         test: /\.scss$/,
         loader: 'style!css!sass?outputStyle=expanded&' +
         'includePaths[]=' +
-        (path.resolve(__dirname, './app/libraries')) + '&' +
+        (path.resolve(__dirname, './app/bower_components')) + '&' +
         'includePaths[]=' +
         (path.resolve(__dirname, './node_modules'))
       },
@@ -43,22 +43,19 @@ module.exports = {
       },
       {
         test: /\.(png|jpg|gif)$/,
-        loader: 'url-loader?limit=8192&name=[name].[ext]'
+        loader: 'file-loader?name=[name].[ext]'
       },
       {
-        test: /\.woff(2)?(\?v=[0-9]\.[0-9]\.[0-9])?$/,
-        loader: 'url-loader?limit=8192&minetype=application/font-woff&name=[name].[ext]'
-      },
-      {
-        test: /\.(ttf|eot|svg)(\?v=[0-9]\.[0-9]\.[0-9])?$/,
+        test: /\.(ttf|eot|svg|woff(2)?)(\?v=[0-9]\.[0-9]\.[0-9])?$/,
         loader: 'file-loader?name=[name].[ext]'
       }
     ]
   },
   plugins: [
     new bowerWebpackPlugin({
-      modulesDirectories: ['app/libraries'],
-      excludes: /.*\.less/
+      modulesDirectories: [path.resolve(__dirname, './app/bower_components')],
+      excludes: /.*\.less/,
+      searchResolveModulesDirectories: false
     }),
     new webpack.ProvidePlugin({
       '$': 'jquery',