+++ app/templates/app/dispatcher/NotifyDispatcher.js
@@ -1,0 +1,4 @@
+var Dispatcher = require('flux').Dispatcher;
+var NotifyDispatcher = new Dispatcher();
+
+module.exports = NotifyDispatcher;